First code for press effects for the scroll line and handle.
This creates a flexible Press/Hover handler that allows you to only apply it to a part of the element (eg for scrollbar subcomponents) and update the area dynamically.

splitRectVert utility function for splitting a rectangle in two.
getSingleElement utility function for getting the element from a single element array, but asserting first.